chore: move icons and projecImage to assets instead of utility
