+++ src/assets/icons.ts
@@ -1,0 +1,47 @@
+import adobeXD from "./logos/adobexd.svg";
+import angular from "./logos/angular.svg";
+import babel from "./logos/babel.svg";
+import css from "./logos/css.svg";
+import docker from "./logos/docker.svg";
+import figma from "./logos/figma.svg";
+import gatsby from "./logos/gatsby.svg";
+import git from "./logos/git.svg";
+import go from "./logos/go.svg";
+import graphql from "./logos/graphql.svg";
+import html5 from "./logos/html5.svg";
+import jest from "./logos/jest.svg";
+import js from "./logos/js.svg";
+import mongodb from "./logos/mongodb.svg";
+import mysql from "./logos/mysql.svg";
+import nodejs from "./logos/nodejs.svg";
+import react from "./logos/react.svg";
+import redux from "./logos/redux.svg";
+import rtl from "./logos/rtl.png";
+import sass from "./logos/sass.svg";
+import ts from "./logos/typescript.svg";
+import webpack from "./logos/webpack.svg";
+
+export const icons = {
+  adobeXD,
+  gatsby,
+  graphql,
+  html5,
+  mysql,
+  nodejs,
+  react,
+  sass,
+  ts,
+  figma,
+  angular,
+  go,
+  js,
+  css,
+  webpack,
+  redux,
+  mongodb,
+  jest,
+  rtl,
+  git,
+  docker,
+  babel,
+};
