+++ src/assets/projectImage.ts
@@ -1,0 +1,17 @@
+import marbond from "./projectImages/project-1.webp";
+import suggestMeAnime from "./projectImages/project-2.webp";
+import pomierskiBlog from "./projectImages/project-3.webp";
+import organize from "./projectImages/project-4.webp";
+import silver from "./projectImages/project-5.webp";
+import myAnimeReminder from "./projectImages/project-6.webp";
+import yayNay from "./projectImages/project-7.webp";
+
+export const projectImage = {
+  marbond,
+  suggestMeAnime,
+  pomierskiBlog,
+  organize,
+  silver,
+  myAnimeReminder,
+  yayNay,
+};
