--- src/utility/icons.ts
@@ -1,47 +1,0 @@
-import adobeXD from "../assets/logos/adobexd.svg";
-import angular from "../assets/logos/angular.svg";
-import babel from "../assets/logos/babel.svg";
-import css from "../assets/logos/css.svg";
-import docker from "../assets/logos/docker.svg";
-import figma from "../assets/logos/figma.svg";
-import gatsby from "../assets/logos/gatsby.svg";
-import git from "../assets/logos/git.svg";
-import go from "../assets/logos/go.svg";
-import graphql from "../assets/logos/graphql.svg";
-import html5 from "../assets/logos/html5.svg";
-import jest from "../assets/logos/jest.svg";
-import js from "../assets/logos/js.svg";
-import mongodb from "../assets/logos/mongodb.svg";
-import mysql from "../assets/logos/mysql.svg";
-import nodejs from "../assets/logos/nodejs.svg";
-import react from "../assets/logos/react.svg";
-import redux from "../assets/logos/redux.svg";
-import rtl from "../assets/logos/rtl.png";
-import sass from "../assets/logos/sass.svg";
-import ts from "../assets/logos/typescript.svg";
-import webpack from "../assets/logos/webpack.svg";
-
-export const icons = {
-  adobeXD,
-  gatsby,
-  graphql,
-  html5,
-  mysql,
-  nodejs,
-  react,
-  sass,
-  ts,
-  figma,
-  angular,
-  go,
-  js,
-  css,
-  webpack,
-  redux,
-  mongodb,
-  jest,
-  rtl,
-  git,
-  docker,
-  babel,
-};
